Area compass

+++ src/main/java/dev/toma/gunsrpg/util/RenderUtils.java
@@ -121,7 +121,7 @@ public class RenderUtils {
         drawLine(pose, x1, y1, x2, y2, color, color);
     }
 
-    public static void drawLine(Matrix4f pose, int x1, int y1, int x2, int y2, int color1, int color2) {
+    public static void drawLine(Matrix4f pose, float x1, float y1, float x2, float y2, int color1, int color2) {
         int a1 = alpha_i(color1);
         int a2 = alpha_i(color2);
         int r1 = red_i(color1);
